Builds version 0x40. Better DMA thru-put, also incorporates ILA debugging probes
This version incorporates many bug fixes and performance enhancements to the DMA driver. It appears to behave as before with a single HLS stream and very much better, but not perfectly with 2 streams.  Packets are still dropped, but at a much lower rate than before.  With two streams, data corruptions have been observed, but this appears to be episodic, in the sense that sometimes it happens and other times, one can run hours without seeing it.  It seems to change states with either a reboot of the RCE or a restart of rceServer.