Checkpoint - logically correct, but does not meet timing
This is just to check-point before making extensive changes which we
feel are necessaryto get encodeN to dataflow without doing a copy
of the data.

This is 'an almost' a working version; it misses meeting the timing
by a factor of 3. (The C/RTL cosim indicated on the order of 200K
cycles for 128 channels. The problem child is the copyN routine;
it either takes too long or takes too many LUTs.

In addition, the code in writeSymsN is structured to dataflow copyN
and encodeN.  What Vivado HLS seems to have produced is copyN running
serially and encodeN + writeN in dataflow.

What one really wants is all three to pipelined.  I tried this once and
while it synthesized, the C/RTL cosimulation hung.

This commited version passed the C-Simulation/Synthesised and  C/RTL
cosimulation with both
          NCHANs = 8
          NCHANS = MODULE_K_NCHANNELS (i.e. 128)